sklearn-kubeflow-trainer/requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
docstring-parser==0.15
enum-compat==0.0.3
google-api-core==2.16.0
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-cloud-build
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
idna==3.6
kfp==2.6.0
kfp-pipeline-spec==0.3.0
kfp-server-api==2.0.5
kubernetes==26.1.0
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
protobuf==4.25.2
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
tabulate==0.9.0
torch-model-archiver==0.9.0
torchserve==0.9.0
urllib3==1.26.18
websocket-client==1.7.0
google-cloud-aiplatform

joblib
pandas
scikit-learn
black
